Saving and Restoring Preferences
Settings class was added to the data library. This class will manage
86BoxLauncher settings.

Settings object is created in the MainWindow and passed to
PreferencesDialog when preferences are opened. The PreferencesDialog
uses a settings object to restore previous settings and write new
settings if the dialog is accepted.

The preferences icon was renamed to be unique, and we can display it for
the preferences button. This change was needed as we now set the
application's custom icon theme as a fallback and use the OS icon theme
as the primary one.

The src directory was added to include directories.