Fixing Icons for Windows
Symbolic links require administrator privileges on Windows and are
problematic because of that. Therefore, icons behind symbolic links
are now copied to the binary directory when building on Windows.

We also set Window icons for MainWindow and PreferencesDialog.

Finally, the preferences dialog was changed to use native directory
separators for emulator binary.