Adding Machine List Model
An MVC static library was created to keep the Model View
Controller-related code. The initial version of the MachineListModel was
added there.

The MainWindows was updated to contain and use the MachineListModel
object on the virtual machines list view. Adding machines already works,
but saving and restoring them is yet to be implemented.

The Machine class was improved to have restore and save methods. Now,
machines also have no icon by default to avoid unnecessary icon loading.

The MachineDialog was improved to update its Machine object with new
values when dialog is accepted.

All unit tests are now in the tests folder and are only built when the
BUILD_TEST is set ON. A new test was added to ensure that the Machinge
class saving and restoring functions are functioning correctly.