Saving and Restoring Machines
The MainWindow was updated with functions to save and restore machines.
Machines are saved into indented JSON files, which are easy to edit if
needed. The MainWindow was further improved by turning on and off
toolbar actions based on whether a machine is selected.

The MachineDialog was updated with a file dialog so users can select new
or existing config files.

The new CMAKE option ENABLE_ASAN can be set to ON for building with the
AddressSanitizer.