Fixing Icons for Qt5
In Qt5 with the Breeze theme, machine icons were replaced with the
generic CPU icon. This problem was fixed by using the system's default
theme as a fallback and making the application's custom theme the
primary theme.

The above change introduced a new problem again in Qt5. The icon
arrow-up and arrow-down icons were replaced with just the arrow icon.
Changing them to go-up and go-down fixed the problem. Now, icons work
nicely on Plasma 6 with the Breeze theme for both Qt5 and Qt6.