List of Other Icons
A list of standard theme icons was added to the asset-src directory so
we can later use this list to provide the required icons from the breeze
theme or similar for Windows and possibly for macOS builds. We expect
icons to be available from the user's system for Linux builds.