Machine Delegate
Machine Deletage was created to display a summary under the name. The
docs folder contains a diagram of the delegate layout and test themes.

While testing the delegate, we discovered that the icon stayed in the
list view after removing it. This bug was fixed by setting a null icon
on Machine::loadIcon when the iconType is Machine::NoIcon.

The MachineListModel was fixed to work with Qt5, where list erase does
not work with const iterators.