Breeze Icons for Windows Version
A few icons from the Breeze theme were added to the project for Windows
users. On Linux systems, the fallback icon theme comes from the user's
desktop environment.

The project was modified for the Windows version to use a breeze or
breeze-dark theme based on the window background and text color values.

For the Windows version, dialog button box icons are now manually set.

Smaller screenshots were added without the need for scaling.