Adding Documentation
This set of changes contains the first steps to creating documentation
of the program's source code. The goal is to make it easier to
understand how the program works and to contribute.

If Doxygen and Graphviz are installed, the `docs` target is added to the
project. Compiling the `docs` target will build documentation to the
docs/html folder in the build directory.