Adding Documentation for GUI Components
This commit mainly adds documentation for GUI components. However, we
also made some minor edits to previously documented code.

We enabled the gui/utilities for all operating systems, and the
setDialogBoxIcons() is now run on all systems for both dialogs. While
redundant for some systems, it only causes a little extra processing.

A new icon was added to the Doxygen documentation as the project logo.
We previously used a 64 x 64 icon, but the Doxygen documentation states
that the maximum height of the logo should not exceed 55 pixels.
Therefore, the new logo is the same image as before, but we have removed
some empty margins to reduce the size to 64 x 54 pixels.