Remember Icon Settings Even if Loading Fails
We tested how the program works if we leave the icon type and name as
they are when the icon loading fails. The program seems to work without
any problems, and the missing icon is shown as empty. Restoring the icon
file causes the icon to reappear the next time the program is run.