Adding Context Menu for Machine List View
A QAction object was created for each tool button, allowing us to use
the same action for buttons and menu items. The createToolButton method
was replaced with a new version that sets up the tool button from the
QAction object and then sets up a few other properties.

A new context menu object was created. It is displayed when the user
requests a context menu from the machine list over a valid item.

16 x 16 size versions of icons were added for actions in the context
menu.